\name{getHistory.TSMLCARA}
\alias{getHistory.TSMLCARA}
\alias{getHistory}
\title{Retrieves the History of a TSMLCARA Object}
\description{Retrieves the \code{history} of a TSMLCARA object.}
\usage{\method{getHistory}{TSMLCARA}(this, ...)}
\arguments{
  \item{this}{A \code{TSMLCARA} object.}
  \item{\dots}{Not used.}
}

\value{The history of the given TSMLCARA object.}

\author{Antoine Chambaz [aut, cre]}



\seealso{as.character}
\examples{FALSE}
