\name{getObs.TSMLCARA}
\alias{getObs.TSMLCARA}
\alias{getObs}
\title{Retrieves the Current Data Set}
\description{Retrieves the \code{data.frame} of observations currently involved in the procedure.}
\usage{\method{getObs}{TSMLCARA}(this, ...)}
\arguments{
  \item{this}{An object of class \code{TSMLCARA}.}
  \item{\dots}{Not used.}
}

\value{The data set involved in  the procedure.}

\author{Antoine Chambaz [aut, cre]}



\seealso{tsml.cara.rct}
\examples{FALSE}
