\name{getPsi.TSMLCARA}
\alias{getPsi.TSMLCARA}
\alias{getPsi}
\title{Returns the Current Estimator}
\description{Returns the current value of the estimator.}
\usage{\method{getPsi}{TSMLCARA}(this, ...)}
\arguments{
  \item{this}{An object of class \code{TSMLCARA}.}
  \item{\dots}{Not used.}
}

\value{Retrieves  the current value  of the estimator. }

\author{Antoine Chambaz [aut, cre]}



\seealso{tsml.lcara.rct, getHistory, getPsiSd}
\examples{FALSE}
