\name{getPsiSd.TSMLCARA}
\alias{getPsiSd.TSMLCARA}
\alias{getPsiSd}
\title{Returns the Current Estimated Standard Deviation of the Current Estimator }
\description{Returns  the current  value of  the  estimated standard  deviation of  the
current estimator.}
\usage{\method{getPsiSd}{TSMLCARA}(this, ...)}
\arguments{
  \item{this}{An object of class \code{TSMLCARA}.}
  \item{\dots}{Not used.}
}

\value{Retrieves the  estimated standard deviation of the  current estimator of
the parameter  of interest.}

\author{Antoine Chambaz [aut, cre]}



\seealso{tsml.lcara.rct, getHistory, getPsi}
\examples{FALSE}
