\name{oneOne}
\alias{oneOne}
\title{Balanced Treatment Mechanism}
\description{Balanced treatment mechanism.}
\usage{oneOne(W)}
\arguments{
  \item{W}{A \code{data.frame} of covariates.}
}

\value{Returns a \code{vector} of probabilities, all equal to 1/2.}

\author{Antoine Chambaz [aut, cre]}





