\name{printHistory}
\alias{printHistory}
\title{Prints a Summary of an History of a TSMLCARA Object}
\description{Prints a summary of an history of a TSMLCARA object.}
\usage{printHistory(x, ...)}
\arguments{
  \item{x}{The \code{history} of a TSMLCARA object.)}
  \item{\dots}{
}
}



\author{Antoine Chambaz [aut, cre]}



\seealso{getHistory}

