\name{ruleQbar}
\alias{ruleQbar}
\title{Computes the Treatment Rule Associated with Qbar}
\description{Computes the Treatment Rule Associated with Qbar}
\usage{ruleQbar(Qbar, ...)}
\arguments{
  \item{Qbar}{A \code{function}, a conditional expectation of Y given (A,W)}
  \item{\dots}{Arguments to be passed to 'Qbar'  }
}

\value{Returns a \code{function}, the rule associated with 'Qbar'.}

\author{Antoine Chambaz [aut, cre]}





