\name{setConfLevel.TSMLCARA}
\alias{setConfLevel.TSMLCARA}
\alias{setConfLevel}
\title{Sets a Confidence Level}
\description{Sets the confidence level of a \code{TSMLCARA} object.}
\usage{\method{setConfLevel}{TSMLCARA}(this, confLevel, ...)}
\arguments{
  \item{this}{An object of class \code{TSMLCARA}.}
  \item{confLevel}{A \code{numeric}, confidence interval level.}
  \item{\dots}{Not used.}
}



\author{Antoine Chambaz [aut, cre]}



\seealso{as.character}
\examples{FALSE}
