\name{targetPsi.TSMLCARA}
\alias{targetPsi.TSMLCARA}
\alias{targetPsi}
\title{Targets a TSMLCARA Object Toward the Parameter Psi}
\description{Targets a TSMLCARA object toward the parameter Psi.}
\usage{\method{targetPsi}{TSMLCARA}(this, ..., verbose = FALSE)}
\arguments{
  \item{this}{A \code{TSMLCARA} object, as created by \code{TSMLCARA}.}
  \item{\dots}{Additional parameters.}
  \item{verbose}{A \code{logical}  or an \code{integer}  indicating the level  of verbosity
(defaults to 'FALSE').}
}


\references{Chambaz, van der Laan,  Zheng, Chapter 16, Modern Adaptive Randomized Clinical  Trials: Statistical, Operational, and Regulatory  Aspects,  by  A.    Sverdlov  (CRC  Press,  2015).}
\author{Antoine Chambaz [aut, cre]}



\seealso{tsml.cara.rct, update}
\examples{FALSE}
